requirements.txt
pandas
nltk
gensim
python-telegram-bot
black
isort
pytest
python-dotenv
